requirements.txt
django==5.1.4
graphene-django==3.2.2
django-graphql-jwt==0.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
freezegun==1.5.1